fixes for so that program compiles error & warning free on linux. Removed all nasty memory access errors due to UChar[] usage. Code completion fixes as well; the code completion now fully works when attempting to trigger completion from the last of a function
The code completion fix involved separating out the scope finding logic into its own class so that it could be called independently from the symbol table creation code

+++ src/plugins/FinderPluginClass.cpp
@@ -215,7 +215,7 @@ mvceditor::ReplacePanelClass::ReplacePanelClass(wxWindow* parent, mvceditor::Not
 		, CurrentInsertionPointReplace(0) {
 	mvceditor::RegularExpressionValidatorClass regExValidator(&Finder.Expression, FinderMode);
 	FindText->SetValidator(regExValidator);
-	wxGenericValidator modeValidator(&(int)Finder.Mode);
+	wxGenericValidator modeValidator(&Finder.Mode);
 	FinderMode->SetValidator(modeValidator);
 	wxGenericValidator wrapValidator(&Finder.Wrap);
 	Wrap->SetValidator(wrapValidator);
